+++ src/types/AccountDetailsMobile.tsx
@@ -1,6 +1,0 @@
-import { TextInput } from 'react-native';
-import { AccountDetailsFormProps } from './AccountDetails';
-
-export type AccountDetailsFormMobileProps = AccountDetailsFormProps & {
-    ref?: React.RefObject<TextInput>;
-};