+++ src/types/index.tsx
@@ -1,5 +1,4 @@
 export * from './AccountDetails';
-export * from './AccountDetailsMobile';
 export * from './ContactParams';
 export * from './ResetPasswordParams';
 export * from './TextInputHTMLAttributes';